cria utils.py com metodos auxiliares gerais
